Add some more examples and restructure
