+++ app/javascript/app.sass
@@ -1,0 +1,2 @@
+.container
+  font-size: 20px
