+++ app/javascript/packs/application.js
@@ -1,13 +1,13 @@
 import Vue from 'vue/dist/vue.esm';
-import App from './app.vue';
+import App from '../app.vue';
 import Vuetify from 'vuetify';
 
 import 'vuetify/dist/vuetify.min.css'
+import '../app.sass'
 
 Vue.use(Vuetify);
 
 document.addEventListener('DOMContentLoaded', () => {
-
 	const app = new Vue({
 		el: '#app',
 		template: '<App/>',
@@ -15,5 +15,4 @@ document.addEventListener('DOMContentLoaded', () => {
 			App
 		}
 	});
-
 });
