+++ app/views/layouts/application.html.erb
@@ -4,9 +4,12 @@
     <title>Ander</title>
     <%= csrf_meta_tags %>
 
-    <%= stylesheet_link_tag    'application', media: 'all', 'data-turbolinks-track': 'reload' %>
-    <%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload' %>
-    <link href='https://fonts.googleapis.com/css?family=Roboto:300,400,500,700|Material+Icons' rel="stylesheet" type="text/css">
+    <link
+      href='https://fonts.googleapis.com/css?family=Roboto:300,400,500,700|Material+Icons'
+      rel="stylesheet"
+      type="text/css"
+    />
+
     <%= javascript_pack_tag 'application' %>
     <%= stylesheet_pack_tag 'application' %>
   </head>
